Limit XP Contribution
Experience bonuses are counted in the formula so that someone opting
to use a Ruby over an Amethyst in their helmet would not be penalized
too greatly.  However, with the addition of the Hellfire ring,
experience bonuses were giving Hellfire ring users a huge bonus.  By
limiting the experience bonus to 35%, the original intent is kept in
tact while preventing Hellfire abuse.